test/pylib: refactor import paths to connect_options module
The suite package is now empty except for python.py, which no longer
holds any suite machinery: it just registers the --host, CQL and S3
connection options. Rename it to test/pylib/connect_options.py to
reflect its actual contents, update the conftest importers accordingly,
and delete the now-empty suite/__init__.py, removing the suite package
for good.
