+++ test/cql/conftest.py
@@ -13,7 +13,7 @@ import pytest
 
 from test.cqlpy.conftest import host, cql, this_dc  # add required fixtures
 from test.pylib.cql_repl import CQL_TEST_SUFFIX, CqlFile
-from test.pylib.suite.python import add_host_option, add_cql_connection_options
+from test.pylib.connect_options import add_host_option, add_cql_connection_options
 
 if TYPE_CHECKING:
     from collections.abc import Generator
