+++ test/cqlpy/conftest.py
@@ -20,7 +20,7 @@ import time
 import random
 
 from test.pylib.skip_types import skip_env
-from test.pylib.suite.python import add_host_option, add_cql_connection_options, add_s3_options
+from test.pylib.connect_options import add_host_option, add_cql_connection_options, add_s3_options
 from test.pylib.scylla_cluster import ScyllaCluster
 from .util import unique_name, new_test_keyspace, keyspace_has_tablets, cql_session, local_process_id, is_scylla, config_value_context
 from .nodetool import scylla_log
