+++ test/pylib/connect_options.py
