+++ test/rest_api/conftest.py
@@ -16,7 +16,7 @@ from test.conftest import dynamic_scope
 from test.cqlpy.conftest import host, cql, this_dc  # add required fixtures
 from test.cqlpy.util import unique_name, new_test_keyspace, keyspace_has_tablets, is_scylla
 from test.pylib.skip_types import skip_env
-from test.pylib.suite.python import add_host_option, add_cql_connection_options
+from test.pylib.connect_options import add_host_option, add_cql_connection_options
 
 # By default, tests run against a Scylla server listening
 # on localhost:9042 for CQL and localhost:10000 for the REST API.
